dash_powerbi/requirements.txt
pandas==2.3.3
dask.dataframe==2023.2.0
